Distances measured in Tile units
I noticed the map vision behaving poorly.
It assumed radius was in tiles, but it was half-tiles.
This broke, so I fixed it naively.
This lead to using 4x as large an array as needed.

Fixing this should help against future mixups.
It is annoying to type the .5, but thats a small price.

+++ Crawler/Controller/AbilityTargetInputState.cs
@@ -39,7 +39,7 @@ public class AbilityTargetInputState : InputState
 
         if (ev.IsActionPressed("ui_accept"))
         {
-            int distance = GridHelper.Distance(playerPos, cursor.targetPosition);
+            float distance = GridHelper.Distance(playerPos, cursor.targetPosition);
             GD.Print(distance);
             if (action.Range.min <= distance && distance <= action.Range.max)
             {
@@ -51,7 +51,7 @@ public class AbilityTargetInputState : InputState
         {
             if (evMouseButton.ButtonIndex == (int)ButtonList.Left && evMouseButton.IsPressed())
             {
-                int distance = GridHelper.Distance(playerPos, cursor.targetPosition);
+                float distance = GridHelper.Distance(playerPos, cursor.targetPosition);
                 GD.Print(distance);
                 if (action.Range.min <= distance && distance <= action.Range.max)
                 {
